requirements.txt
pillow
pdfminer.six
chardet==3.0.4
pyocr
pytesseract
pandas
python-dotenv
pyyaml>=6.0
pdf2image==1.17.0
poppler-utils==0.1.0